+++ wayland-protocols/src/staging.rs
@@ -41,6 +41,28 @@ pub mod xdg_activation {
     }
 }
 
+pub mod ext_session_lock {
+    //! This protocol allows for a privileged Wayland client to lock the session
+    //! and display arbitrary graphics while the session is locked.
+    //!
+    //! The compositor may choose to restrict this protocol to a special client
+    //! launched by the compositor itself or expose it to all privileged clients,
+    //! this is compositor policy.
+    //!
+    //! The client is responsible for performing authentication and informing the
+    //! compositor when the session should be unlocked. If the client dies while
+    //! the session is locked the session remains locked, possibly permanently
+    //! depending on compositor policy.
+
+    #[allow(missing_docs)]
+    pub mod v1 {
+        wayland_protocol!(
+            "./protocols/staging/ext-session-lock/ext-session-lock-v1.xml",
+            []
+        );
+    }
+}
+
 pub mod drm_lease {
     //! This protocol is used by Wayland compositors which act as Direct
     //! Renderering Manager (DRM) masters to lease DRM resources to Wayland